+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.2.0
-* Added `isReady` getter to `ILibJS` class to expose internal `_iLibPrepared` state.
-* Added `isReady` getter to `FlutterILib` class for external access to ILibJS initialization status.
+* Added `isILibReady` getter to `ILibJS` class to expose internal `_iLibPrepared` state.
+* Added `isILibReady` getter to `FlutterILib` class for external access to ILibJS initialization status.
 
 ## 1.1.0
 * Updated the iLib files to version 14.21.0 since the new version of iLib has been released.